pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "conda-pypi"
description = "Better PyPI interoperability for the conda ecosystem."
readme = "README.md"
authors = [
  { name = "Jaime Rodríguez-Guerra", email = "jrodriguez@quansight.com" },
]
license = "MIT"
license-files = [
  "LICENSE",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
requires-python = ">=3.10"
dependencies = [
  # "conda >=25.11.0",
  "pip",
  "build",
  "conda-index",
  "installer",
  "packaging",
  "platformdirs",
  "unearth",
]
dynamic = ["version"]

[project.urls]
homepage = "https://github.com/conda/conda-pypi"

[project.entry-points.conda]
conda-pypi = "conda_pypi.plugin"

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64", "osx-64", "osx-arm64", "win-64", "linux-aarch64"]

[tool.pixi.dependencies]
# must match project.dependencies (sans python and conda)
python = ">=3.10"
conda = ">=26.1"
pip = "*"
packaging = "*"
conda-index = ">=0.7.0"
pytest-codspeed = ">=3.0.0,<5"
conda-rattler-solver = ">=0.0.5"

[tool.pixi.pypi-dependencies]
"conda-pypi" = { path  = ".", editable = true }

[tool.pixi.tasks]
dev = 'python -c "from conda_pypi.python_paths import ensure_externally_managed as e; e()"'

[tool.pixi.feature.build]
dependencies = { anaconda-client = "*", conda-build = "*" }
tasks = { build = "conda build recipe" }

[tool.pixi.feature.lint.dependencies]
python = "3.12.*"
pre-commit = "*"

[tool.pixi.feature.lint.tasks]
lint = "pre-commit"
pre-commit = "pre-commit"

[tool.pixi.feature.docs.tasks]
docs = { cmd = "sphinx-build -M dirhtml . _build", cwd = "docs" }
serve = { cmd = "python -m http.server", cwd = "docs/_build/dirhtml" }
clean = { cmd = "rm -rf _build", cwd = "docs" }

[tool.pixi.feature.docs.dependencies]
python = "3.10.*"
conda-sphinx-theme = "*"
linkify-it-py = "*"
myst-parser = "*"
sphinx = "*"
sphinx-argparse = "*"
sphinx-copybutton = "*"
sphinx-design = "*"
sphinx-reredirects = "*"
sphinx-sitemap = "*"

[tool.pixi.feature.test.tasks]
test = 'python -c "from conda_pypi.python_paths import ensure_externally_managed as e; e()" && python -mpytest -m "not benchmark"'
benchmark= 'python -c "from conda_pypi.python_paths import ensure_externally_managed as e; e()" && python -mpytest --codspeed'

[tool.pixi.feature.test.dependencies]
pytest = "7.4.3.*"
fmt = "!=10.2.0"
pytest-mock = "3.12.0.*"
pytest-xprocess = "*"
conda-build = "*"
# Build system dependencies
conda-package-streaming = ">=0.11"
unearth = ">=0.17.2"
python-build = "*"
python-installer = ">=0.7"
packaging = ">=24"
# Dependencies for corpus tests
sqlalchemy = "*"
python-multipart = "*"
pyyaml = "*"
zstandard = "*"  # Will be in stdlib as compression.zstd in Python 3.14+
# Build backend for flit tests
flit = "*"

[tool.pixi.feature.py310.dependencies]
python = "3.10.*"

[tool.pixi.feature.py311.dependencies]
python = "3.11.*"

[tool.pixi.feature.py312.dependencies]
python = "3.12.*"

[tool.pixi.feature.py313.dependencies]
python = "3.13.*"

[tool.pixi.feature.py314.dependencies]
python = "3.14.*"

[tool.pixi.feature.canary]
channels = ["conda-forge", {channel = "conda-canary/label/dev", priority = -1}]

[tool.pixi.feature.canary.dependencies]
conda = { version = ">=26.1", channel = "conda-canary/label/dev" }

[tool.pixi.environments]
build-py310 = ["build", "py310"]
build-py311 = ["build", "py311"]
build-py312 = ["build", "py312"]
build-py313 = ["build", "py313"]
build-py314 = ["build", "py314"]
docs = ["docs"]
lint = ["lint"]
test-py310 = ["test", "py310"]
test-py311 = ["test", "py311"]
test-py312 = ["test", "py312"]
test-py313 = ["test", "py313"]
test-py314 = ["test", "py314"]
test-canary = ["test", "py313", "canary"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "conda_pypi/_version.py"

[tool.ruff]
line-length = 99

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]

[tool.coverage.run]
source = ["conda_pypi/", "tests/"]
omit = ["conda_pypi/__init__.py"]
